+++ apps/web/utils/Constants.ts
@@ -87,3 +87,9 @@ export const maxReceiptCharacters = 200;
 export const maxDescriptionCharacters = 500;
 export const maxLogoNameCharacters = 100;
 export const HTTP_STATUS_UNAUTHORIZED = 401;
+export const CARD_BRANDS = {
+  VISA: "visa",
+  MASTERCARD: "mastercard",
+} as const;
+
+export type CardBrand = (typeof CARD_BRANDS)[keyof typeof CARD_BRANDS];
